修改柯莱的阵营
